feat: add watcher for events
Problem: I did not see a nice way to get (from the scheduler)
a record of when a pod was done on a node. Instead, it uses
the events API to go off of updated snapshots.
Solution: Do the same! We can get pretty good metadata
for pods and nodes via the same events API.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>